// cgocore-ui/src/app/features/models/reference/airlines/views/basic/AirlineCompanyBasicView.ts
/**
 * auto generated basic view bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class AirlineCompanyBasicView {

id: number;
selected: boolean;
canDelete: boolean;
iataCode: string;
label: string;
prefix: string;

/* Specific Code Start */
/* Specific Code End */
}