// cgocore-ui/src/app/features/models/reference/airlines/sortings/AirlineCompanySorting.ts
/**
 * auto generated sorting ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class AirlineCompanySorting {

iataCodeOrderType: string;
labelOrderType: string;
prefixOrderType: string;

/* Specific Code Start */
/* Specific Code End */
}